requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiogram==3.10.0
pydantic>=2.4.1,<2.9.0
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
asyncpg>=0.29.0
alembic==1.13.2
httpx==0.27.2
python-dotenv==1.0.1
loguru==0.7.2
jinja2==3.1.4
pyyaml==6.0.2
python-multipart==0.0.6
greenlet>=3.0.0
